add CRUD controlers
Comments is per convention the CRUD controler for Comment.
The convention is to pluralize the model object name to create the corresponding controller.
This way, Play will automatically find the associated model object for each controller.
If you need to use a different name, you can still use the @CRUD.For annotation.

The CRUD module is accessible under:
http://localhost:9000/admin/