Add a captcha to avoid comments' spam
To validate the captcha we need to save the random text written to the captcha
image somewhere and then check it at the form submission time.

Of course we could just put the text to the user session at the image generation
time and then retrieve it later. But this solution has two drawbacks:

First, the Play session is stored as a cookie. It solves a lot of problems in terms of
architecture but has a lot of implications. Data written to the session cookie are
signed (so the user can’t modify them) but not encrypted. If we write the captcha code
to the session anybody could easily resolve it by reading the session cookie.

Second, remember that Play is a stateless framework. We want to manage things in a purely
stateless way. Typically, what happens if a user simultaneously opens two different
blog pages with two different captcha images? We have to track the captcha code for each form.

So to resolve the problem we need two things. We will store the captcha secret key on the
server side. Because it is transient data we can easily use the Play Cache. Moreover because
cached data have a limited life time it will add one more security mechanism (let’s say that
a captcha code will be available for only 10mn). Then to resolve the code later we need to
generate a unique ID. This unique ID will be added to each form as an hidden field and
implicitly references a generated captcha code.